Update to 0.18.1696 rev1102 (development build)

+++ Assembly-CSharp/RimWorld/PawnGenOption.cs
@@ -7,7 +7,7 @@ namespace RimWorld
 	{
 		public PawnKindDef kind;
 
-		public int selectionWeight;
+		public float selectionWeight;
 
 		public float Cost
 		{
@@ -25,7 +25,7 @@ namespace RimWorld
 		public void LoadDataFromXmlCustom(XmlNode xmlRoot)
 		{
 			DirectXmlCrossRefLoader.RegisterObjectWantsCrossRef(this, "kind", xmlRoot.Name);
-			this.selectionWeight = (int)ParseHelper.FromString(xmlRoot.FirstChild.Value, typeof(int));
+			this.selectionWeight = (float)ParseHelper.FromString(xmlRoot.FirstChild.Value, typeof(float));
 		}
 	}
 }
